nats_app/requirements.txt
nats-py==2.1.0
python-dotenv==1.0.0
